# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Collect&Verything",
        "version": "0.1"
    },
    "paths": {
...
        "/api/stores/{id}/closings": {
            "post": {
                "tags": [
                    "Stores Closings"
                ],
                "summary": "Post a new store closing",
                "description": "Create a closing",
                "operationId": "addClosing",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Store Id",
                        "required": true
                    },
                    {
                        "name": "from",
                        "in": "query",
                        "description": "From",
                        "required": true
                    },
                    {
                        "name": "to",
                        "in": "query",
                        "description": "To",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Closing created"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/stores/{id}/closings/{closing_id}": {
            "delete": {
                "tags": [
                    "Stores Closings"
                ],
                "summary": "Delete a store closing",
                "description": "Soft delete a closing",
                "operationId": "removeClosing",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Store id",
                        "required": true
                    },
                    {
                        "name": "closing_id",
                        "in": "query",
                        "description": "Closing id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Closing deleted"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
        "/api/stores/{id}/images": {
            "post": {
                "tags": [
                    "Stores Images"
                ],
                "summary": "Post a new store image",
                "description": "Create a image",
                "operationId": "addImage",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Store Id",
                        "required": true
                    },
                    {
                        "name": "image_id",
                        "in": "query",
                        "description": "Image Id",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Image created"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/stores/{id}/images/{store_image_id}": {
            "delete": {
                "tags": [
                    "Stores Images"
                ],
                "summary": "Delete a store image",
                "description": "Soft delete a store image",
                "operationId": "removeImage",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Store id",
                        "required": true
                    },
                    {
                        "name": "store_image_id",
                        "in": "query",
                        "description": "Image id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Image deleted"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
        "/api/stores/{id}/slots": {
            "post": {
                "tags": [
                    "Stores Slots"
                ],
                "summary": "Post a new store slot",
                "description": "Create a slot",
                "operationId": "addSlot",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Store Id",
                        "required": true
                    },
                    {
                        "name": "day",
                        "in": "query",
                        "description": "Day",
                        "required": true
                    },
                    {
                        "name": "from",
                        "in": "query",
                        "description": "From",
                        "required": true
                    },
                    {
                        "name": "to",
                        "in": "query",
                        "description": "To"
                    },
                    {
                        "name": "quantity",
                        "in": "query",
                        "description": "Quantity",
                        "required": true
                    },
                    {
                        "name": "available",
                        "in": "query",
                        "description": "Availability",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Account created"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/stores/{id}/slots/{slot_id}": {
            "delete": {
                "tags": [
                    "Stores Slots"
                ],
                "summary": "Delete a store slot",
                "description": "Soft delete a store slot",
                "operationId": "removeSlot",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Store id",
                        "required": true
                    },
                    {
                        "name": "slot_id",
                        "in": "query",
                        "description": "Slot id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Slot deleted"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Stores Slots"
                ],
                "summary": "Update a store slot",
                "description": "update the availability and the quantity of a store slot",
                "operationId": "updateSlot",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Store id",
                        "required": true
                    },
                    {
                        "name": "slot_id",
                        "in": "query",
                        "description": "Slot id",
                        "required": true
                    },
                    {
                        "name": "quantity",
                        "in": "query",
                        "description": "Quantity",
                        "required": false
                    },
                    {
                        "name": "available",
                        "in": "query",
                        "description": "Availability",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Slot updated"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        []
    ]
}